fitness-dashboard/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Google Drive integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Date/time handling
python-dateutil>=2.8.2

# JSON schema validation
jsonschema>=4.19.0

# Rate limiting utilities
ratelimit>=2.2.1

# For Strava OAuth (Phase 2)
# stravalib>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Linting (optional)
pylint>=3.0.0
black>=23.0.0
